+++ components/molecules/ErrorBox.js
@@ -8,7 +8,7 @@ export function ErrorBox(props) {
   return (
     <div
       id="error-box"
-      className="relative border-l-4 border-error-border-red min-h-40px mb-10"
+      className="relative border-l-4 border-error-border-red min-h-40px my-10"
       data-cy="error-box"
       role="alert"
       aria-atomic="true"