+++ components/molecules/Menu.js
@@ -1,6 +1,7 @@
 import PropTypes from "prop-types";
 import Link from "next/link";
 import { useRouter } from "next/router";
+import { useState } from "react";
 
 /**
  * Menu component
@@ -8,17 +9,8 @@ import { useRouter } from "next/router";
 export function Menu(props) {
   //Router
   const { asPath } = useRouter();
-
-  //Function for changing menu state
-  function onMenuClick() {
-    const menuButton = document.getElementById("menuButton");
-    const menuDropdown = document.getElementById("menuDropdown");
-
-    menuDropdown.classList.toggle("active");
-    menuButton.getAttribute("aria-expanded") === "true"
-      ? menuButton.setAttribute("aria-expanded", false)
-      : menuButton.setAttribute("aria-expanded", true);
-  }
+  const router = useRouter();
+  const [showMenu, setShowMenu] = useState(false);
 
   return (
     <nav
@@ -31,22 +23,43 @@ export function Menu(props) {
       <h3 className="sr-only" id="mainSiteNav">
         Menu
       </h3>
-      <button
-        id="menuButton"
-        onClick={onMenuClick}
-        className="text-h4 text-canada-footer-font focus:outline-none focus:ring-2 focus:ring-black mb-4 py-1"
-        aria-haspopup="true"
-        aria-expanded="false"
-        aria-controls="menuDropdown"
-        data-testid="menuButton"
-      >
-        <span className="inline-block align-middle icon-menu" />
-        <span className="inline-block align-middle pl-3 font-body text-p leading-none">
-          {props.menuButtonTitle}
-        </span>
-      </button>
+      <div className="flex justify-between">
+        <button
+          id="menuButton"
+          onClick={() => setShowMenu(!showMenu)}
+          className="text-h4 text-canada-footer-font focus:outline-none focus:ring-2 focus:ring-black mb-4 py-1"
+          aria-haspopup="true"
+          aria-expanded={showMenu}
+          aria-controls="menuDropdown"
+          data-testid="menuButton"
+        >
+          <span className="inline-block align-middle icon-menu" />
+          <span className="inline-block align-middle pl-3 font-body text-p leading-none">
+            {props.menuButtonTitle}
+          </span>
+        </button>
 
-      <ul id="menuDropdown" className="menuDropdown" role="menu">
+        <button
+          id="menuClose"
+          onClick={() => setShowMenu(!showMenu)}
+          className={`${
+            showMenu ? "" : "hidden"
+          } sr-only mb-4 text-canada-footer-font outline-none focus:not-sr-only focus:outline-black-solid lg:invisible`}
+          aria-expanded={showMenu}
+          aria-controls="menuDropdown"
+          aria-label="Close the expanded menu options"
+          data-testid="menuCloseButton"
+        >
+          <img src="/close-x-menu.svg" alt="Close button"></img>
+        </button>
+      </div>
+
+      <ul
+        id="menuDropdown"
+        className={`menuDropdown ${showMenu ? "active" : ""}`}
+        role="menu"
+        aria-expanded={showMenu}
+      >
         {props.items.map((item, key) => {
           const exactURL = asPath === item.link; // it's exactly this url
           const includesURL = asPath.includes(item.link); // it's a child of this url (eg, "/projects/app" includes "/projects")
@@ -61,7 +74,11 @@ export function Menu(props) {
               <Link href={item.link}>
                 <a
                   className={`font-body text-base ${
-                    includesURL ? "activePage" : "menuLink underline"
+                    includesURL
+                      ? router.pathname !== "/signup/privacy"
+                        ? "activePage"
+                        : "menuLink underline"
+                      : "menuLink underline"
                   }`}
                 >
                   {item.text}