+++ pages/404.js
@@ -25,10 +25,19 @@ export default function error404(props) {
           ) : (
             ""
           )}
+
+          {/* Primary HTML Meta Tags */}
           <title data-gc-analytics-error="404">
             {`404 — ${t("siteTitle")}`}
           </title>
+          <meta
+            name="description"
+            content={`${t("404errorMetaDescription")}`}
+          />
+          <meta name="author" content="Service Canada" />
           <link rel="icon" href="/favicon.ico" />
+
+          {/* DCMI Meta Tags */}
           <meta name="dcterms.title" content={`404 — ${t("siteTitle")}`} />
           <meta
             name="dcterms.language"
@@ -41,6 +50,40 @@ export default function error404(props) {
             content="ESDC-EDSC_SCLabs-LaboratoireSC"
           />
           <meta name="dcterms.issued" content="2021-06-01" />
+
+          {/* Open Graph / Facebook */}
+          <meta property="og:type" content="website" />
+          <meta property="og:locale" content={props.locale} />
+          <meta
+            property="og:url"
+            content={
+              "https://alpha.service.canada.ca/" + `${props.locale}` + "/404"
+            }
+          />
+          <meta property="og:title" content={`404 — ${t("siteTitle")}`} />
+          <meta
+            property="og:description"
+            content={`${t("404errorMetaDescription")}`}
+          />
+          <meta property="og:image" content={`${t("metaImage")}`} />
+          <meta property="og:image:alt" content={`${t("siteTitle")}`} />
+
+          {/* Twitter */}
+          <meta property="twitter:card" content="summary_large_image" />
+          <meta
+            property="twitter:url"
+            content={
+              "https://alpha.service.canada.ca/" + `${props.locale}` + "/404"
+            }
+          />
+          <meta property="twitter:title" content={`500 — ${t("siteTitle")}`} />
+          <meta name="twitter:creator" content={t("creator")} />
+          <meta
+            property="twitter:description"
+            content={`${t("404errorMetaDescription")}`}
+          />
+          <meta property="twitter:image" content={`${t("metaImage")}`} />
+          <meta property="twitter:image:alt" content={`${t("siteTitle")}`} />
         </Head>
         <section className="layout-container pb-44">
           <img