+++ pages/home.js
@@ -2,7 +2,6 @@ import Head from "next/head";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { useTranslation } from "next-i18next";
 import { Layout } from "../components/organisms/Layout";
-import { useRouter } from "next/router";
 import { CallToAction } from "../components/molecules/CallToAction";
 import { ActionButton } from "../components/atoms/ActionButton";
 import { HTMList } from "../components/atoms/HTMList";
@@ -10,7 +9,6 @@ import { useEffect } from "react";
 
 export default function Home(props) {
   const { t } = useTranslation("common");
-  const { asPath } = useRouter();
 
   useEffect(() => {
     if (process.env.NEXT_PUBLIC_ADOBE_ANALYTICS_URL) {
@@ -25,7 +23,7 @@ export default function Home(props) {
         bannerTitle={t("siteTitle")}
         bannerText={t("bannerText")}
         locale={props.locale}
-        langUrl={asPath}
+        langUrl={t("homePath")}
       >
         <Head>
           {process.env.NEXT_PUBLIC_ADOBE_ANALYTICS_URL ? (
@@ -33,8 +31,14 @@ export default function Home(props) {
           ) : (
             ""
           )}
+
+          {/* Primary HTML Meta Tags */}
           <title>{`${t("scLabsHome")} — ${t("siteTitle")}`}</title>
+          <meta name="description" content={`${t("homeMetaDescription")}`} />
+          <meta name="author" content="Service Canada" />
           <link rel="icon" href="/favicon.ico" />
+
+          {/* DCMI Meta Tags */}
           <meta
             name="dcterms.title"
             content={`${t("scLabsHome")} — ${t("siteTitle")}`}
@@ -50,11 +54,55 @@ export default function Home(props) {
             content="ESDC-EDSC_SCLabs-LaboratoireSC"
           />
           <meta name="dcterms.issued" content="2021-03-18" />
+
+          {/* Open Graph / Facebook */}
+          <meta property="og:type" content="website" />
+          <meta property="og:locale" content={props.locale} />
+          <meta
+            property="og:url"
+            content={
+              "https://alpha.service.canada.ca/" +
+              `${props.locale}` +
+              `${t("homeMetaPath")}`
+            }
+          />
+          <meta
+            property="og:title"
+            content={`${t("scLabsHome")} — ${t("siteTitle")}`}
+          />
+          <meta
+            property="og:description"
+            content={`${t("homeMetaDescription")}`}
+          />
+          <meta property="og:image" content={`${t("metaImage")}`} />
+          <meta property="og:image:alt" content={`${t("siteTitle")}`} />
+
+          {/* Twitter */}
+          <meta property="twitter:card" content="summary_large_image" />
+          <meta
+            property="twitter:url"
+            content={
+              "https://alpha.service.canada.ca/" +
+              `${props.locale}` +
+              `${t("homeMetaPath")}`
+            }
+          />
+          <meta
+            property="twitter:title"
+            content={`${t("scLabsHome")} — ${t("siteTitle")}`}
+          />
+          <meta name="twitter:creator" content={t("creator")} />
+          <meta
+            property="twitter:description"
+            content={`${t("homeMetaDescription")}`}
+          />
+          <meta property="twitter:image" content={`${t("metaImage")}`} />
+          <meta property="twitter:image:alt" content={`${t("siteTitle")}`} />
         </Head>
         <section className="layout-container my-12">
           <div className="xl:w-2/3">
             <ActionButton
-              href="/signup"
+              href={t("signupRedirect")}
               id="signup-home-page"
               dataCy="signup-home-page"
               className="rounded px-6 py-4 font-bold text-center inline-block"
@@ -63,26 +111,28 @@ export default function Home(props) {
             </ActionButton>
             <h2 className="my-10">{t("projectsAndExplorationTitle")}</h2>
             <p className="mb-4 whitespace-pre-line">
-              {t("experimentsAndExploration-1/3")}
+              {t("experimentsAndExploration-1/4")}
             </p>
+            <p className="mb-4">{t("experimentsAndExploration-2/4")}</p>
             <HTMList
               listClassName={"mb-4 pl-10 text-p list-disc"}
               content={t("experimentsAndExplorationList")}
             />
-            <p className="mb-4">{t("experimentsAndExploration-2/3")}</p>
-            <p className="mb-10">{t("experimentsAndExploration-3/3")}</p>
+            <p className="mb-4">{t("experimentsAndExploration-3/4")}</p>
+            <p className="mb-10">{t("experimentsAndExploration-4/4")}</p>
+            <p className="mb-10">{t("projectsDisclaimerBody")}</p>
           </div>
           <div className="flex flex-col gap-6 lg:gap-10 lg:flex-row ">
             <ActionButton
-              href="/projects"
-              text={t("experimentsBtnTxt")}
+              href={t("projectRedirect")}
+              text={t("menuLink1")}
               id="ProjectsButton"
               dataCy="ProjectsButton"
               className="flex py-2 px-6 justify-center font-bold rounded"
               secondary
             />
             <ActionButton
-              href="/about"
+              href={t("aboutRedirect")}
               text={t("learnMoreAboutSCL")}
               id="AboutButton"
               dataCy="AboutButton"
@@ -94,7 +144,8 @@ export default function Home(props) {
         <CallToAction
           title={t("signupTitleCallToAction")}
           html={t("becomeAParticipantDescription")}
-          href={"/signup"}
+          lang={props.locale}
+          href={t("signupRedirect")}
           hrefText={t("signupBtn")}
         />
       </Layout>