+++ pages/index.js
@@ -31,8 +31,17 @@ export default function Index(props) {
         ) : (
           ""
         )}
+
+        {/* Primary HTML Meta Tags */}
         <title>Service Canada Labs | Laboratoires de Service Canada</title>
+        <meta
+          name="description"
+          content={`${t("scLabsSplash")} — ${t("siteTitle")}`}
+        />
+        <meta name="author" content="Service Canada" />
         <link rel="icon" href="/favicon.ico" />
+
+        {/* DCMI Meta Tags */}
         <meta
           name="dcterms.title"
           content={`${t("scLabsSplash")} — ${t("siteTitle")}`}
@@ -45,6 +54,39 @@ export default function Index(props) {
         <meta name="dcterms.accessRights" content="2" />
         <meta name="dcterms.service" content="ESDC-EDSC_SCLabs-LaboratoireSC" />
         <meta name="dcterms.issued" content="2021-05-06" />
+
+        {/* Open Graph / Facebook */}
+        <meta property="og:type" content="website" />
+        <meta property="og:locale" content={props.locale} />
+        <meta property="og:url" content={"https://alpha.service.canada.ca/"} />
+        <meta
+          property="og:title"
+          content={`${t("scLabsSplash")} — ${t("siteTitle")}`}
+        />
+        <meta
+          property="og:description"
+          content={`${t("scLabsSplash")} — ${t("siteTitle")}`}
+        />
+        <meta property="og:image" content={`${t("metaImage")}`} />
+        <meta property="og:image:alt" content={`${t("siteTitle")}`} />
+
+        {/* Twitter */}
+        <meta property="twitter:card" content="summary_large_image" />
+        <meta
+          property="twitter:url"
+          content={"https://alpha.service.canada.ca/"}
+        />
+        <meta
+          property="twitter:title"
+          content={`${t("scLabsSplash")} — ${t("siteTitle")}`}
+        />
+        <meta name="twitter:creator" content={t("creator")} />
+        <meta
+          property="twitter:description"
+          content={`${t("scLabsSplash")} — ${t("siteTitle")}`}
+        />
+        <meta property="twitter:image" content={`${t("metaImage")}`} />
+        <meta property="twitter:image:alt" content={`${t("siteTitle")}`} />
       </Head>
       <main>
         <div className="flex flex-col justify-center items-center m-auto v-xxs:h-screen">
@@ -79,7 +121,7 @@ export default function Index(props) {
                 id="french-button"
                 text="Français"
                 className="text-center w-7.5rem xl:w-138px text-sm py-3.5 rounded leading-3"
-                href="/fr/home"
+                href="/fr/accueil"
                 lang="fr"
                 onClick={setLanguage}
               />